move auth files to auth folder
