+++ client/src/components/InputsModal/__tests__/Inputs.test.tsx
@@ -7,8 +7,8 @@ import InputCheckboxGroup from '~/components/InputsModal/InputCheckboxGroup';
 import InputRadioGroup from '~/components/InputsModal/InputRadioGroup';
 import InputTextField from '~/components/InputsModal/InputTextField';
 import InputOAuthCredentials from '~/components/InputsModal/InputOAuthCredentials';
-import InputAuth from '~/components/InputsModal/InputAuth';
-import InputAccess from '~/components/InputsModal/InputAccess';
+import InputAuth from '~/components/InputsModal/Auth/InputAuth';
+import InputAccess from '~/components/InputsModal/Auth/InputAccess';
 import { describe, expect, it } from 'vitest';
 
 describe('Input Components', () => {
